# contracts/generated/openapi/pollen.v1.yaml
openapi: 3.1.0
info:
  title: Pollen DEK Control-Plane API
  version: 1.0.0
  contact:
    name: Pollen Platform Team
    url: https://your-cloud-domain.com
    email: platform@your-cloud-domain.com
  description: Contract between Pollen Cloud / Local Control Plane and the DEK edge enforcement plane.
...
paths:
...
  /v1/tenants/{tenant_id}/usage/agents/{agent_id}:
    get:
      operationId: AiUsageApi_getAgentUsageSummary
      description: Get canonical AI usage summary for one agent
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.from'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.to'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.bucket'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.group_by'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.agent_type'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.provider'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.model'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.task_id'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.session_id'
        - $ref: '#/components/parameters/AiUsageAgentSummaryQueryParams.surface'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageSummaryV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
  /v1/tenants/{tenant_id}/usage/budgets:
    get:
      operationId: AiUsageApi_listUsageBudgets
      description: List AI budget policies
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiBudgetLimitListV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
  /v1/tenants/{tenant_id}/usage/budgets/{budget_id}:
    put:
      operationId: AiUsageApi_upsertUsageBudget
      description: Upsert an AI budget policy
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - name: budget_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/AiBudgetLimitV1'
                required:
                  - item
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiBudgetLimitV1'
  /v1/tenants/{tenant_id}/usage/events:
    post:
      operationId: AiUsageApi_ingestUsageEvent
      description: Ingest one canonical AI usage event
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/AiUsageEventV1'
                required:
                  - item
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiUsageEventV1'
    get:
      operationId: AiUsageApi_listUsageEvents
      description: List canonical AI usage events
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
        - $ref: '#/components/parameters/AiUsageQueryParams.from'
        - $ref: '#/components/parameters/AiUsageQueryParams.to'
        - $ref: '#/components/parameters/AiUsageQueryParams.agent_id'
        - $ref: '#/components/parameters/AiUsageQueryParams.agent_type'
        - $ref: '#/components/parameters/AiUsageQueryParams.provider'
        - $ref: '#/components/parameters/AiUsageQueryParams.model'
        - $ref: '#/components/parameters/AiUsageQueryParams.task_id'
        - $ref: '#/components/parameters/AiUsageQueryParams.session_id'
        - $ref: '#/components/parameters/AiUsageQueryParams.surface'
        - $ref: '#/components/parameters/AiUsageQueryParams.sync_status'
        - $ref: '#/components/parameters/AiUsageQueryParams.limit'
        - $ref: '#/components/parameters/AiUsageQueryParams.cursor'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageEventPageV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
  /v1/tenants/{tenant_id}/usage/events/batch:
    post:
      operationId: AiUsageApi_ingestUsageEventsBatch
      description: Ingest canonical AI usage events in batch
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageBatchIngestResponseV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiUsageBatchIngestRequestV1'
  /v1/tenants/{tenant_id}/usage/ledger:
    get:
      operationId: AiUsageApi_getUsageLedger
      description: List canonical AI usage ledger events
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
        - $ref: '#/components/parameters/AiUsageQueryParams.from'
        - $ref: '#/components/parameters/AiUsageQueryParams.to'
        - $ref: '#/components/parameters/AiUsageQueryParams.agent_id'
        - $ref: '#/components/parameters/AiUsageQueryParams.agent_type'
        - $ref: '#/components/parameters/AiUsageQueryParams.provider'
        - $ref: '#/components/parameters/AiUsageQueryParams.model'
        - $ref: '#/components/parameters/AiUsageQueryParams.task_id'
        - $ref: '#/components/parameters/AiUsageQueryParams.session_id'
        - $ref: '#/components/parameters/AiUsageQueryParams.surface'
        - $ref: '#/components/parameters/AiUsageQueryParams.sync_status'
        - $ref: '#/components/parameters/AiUsageQueryParams.limit'
        - $ref: '#/components/parameters/AiUsageQueryParams.cursor'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageLedgerV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
  /v1/tenants/{tenant_id}/usage/reconcile:
    post:
      operationId: AiUsageApi_reconcileUsage
      description: Accept provider billing reconciliation metadata
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageReconcileResponseV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
      requestBody:
        required: true
        content:
          application/json:
            schema: {}
  /v1/tenants/{tenant_id}/usage/summary:
    get:
      operationId: AiUsageApi_getUsageSummary
      description: Get canonical AI usage summary
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PollenHeaders.contractVersion'
        - $ref: '#/components/parameters/PollenHeaders.deviceId'
        - $ref: '#/components/parameters/PollenHeaders.tenantId'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.from'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.to'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.bucket'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.group_by'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.agent_id'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.agent_type'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.provider'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.model'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.task_id'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.session_id'
        - $ref: '#/components/parameters/AiUsageSummaryQueryParams.surface'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiUsageSummaryV1'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PollenError'
      tags:
        - AI Usage
components:
  parameters:
    AiUsageAgentSummaryQueryParams.agent_type:
      name: agent_type
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.bucket:
      name: bucket
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.from:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      explode: false
    AiUsageAgentSummaryQueryParams.group_by:
      name: group_by
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.model:
      name: model
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.provider:
      name: provider
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.session_id:
      name: session_id
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.surface:
      name: surface
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.task_id:
      name: task_id
      in: query
      required: false
      schema:
        type: string
      explode: false
    AiUsageAgentSummaryQueryParams.to:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
      explode: false
...
    AiUsageQueryParams.from:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
      explode: false
...
    PollenHeaders.contractVersion:
      name: X-Pollen-Contract-Version
      in: header
      required: true
      schema:
        type: string
    PollenHeaders.deviceId:
      name: X-Pollen-Device-Id
      in: header
      required: false
      schema:
        type: string
    PollenHeaders.tenantId:
      name: X-Pollen-Tenant-Id
      in: header
      required: false
      schema:
        type: string
  schemas:
...
    AiBudgetLimitListV1:
      type: object
      required:
        - schema_version
        - items
      properties:
        schema_version:
          type: string
          enum:
            - ai-budget-limit-list.v1
        items:
          type: array
          items:
            $ref: '#/components/schemas/AiBudgetLimitV1'
    AiBudgetLimitV1:
      type: object
      required:
        - schema_version
        - budget_id
        - tenant_id
        - scope_type
        - scope_id
        - window
        - currency
        - action_on_soft
        - action_on_hard
        - enabled
        - created_at
        - updated_at
      properties:
        schema_version:
          type: string
          enum:
            - ai-budget-limit.v1
        budget_id:
          type: string
        tenant_id:
          type: string
        scope_type:
          type: string
        scope_id:
          type: string
        window:
          type: string
        currency:
          type: string
        soft_cost_limit:
          type: number
          format: float
        hard_cost_limit:
          type: number
          format: float
        soft_token_limit:
          type: integer
          format: int64
        hard_token_limit:
          type: integer
          format: int64
        action_on_soft:
          type: string
        action_on_hard:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
...
    AiUsageSummaryV1:
      type: object
      required:
        - schema_version
        - tenant_id
        - bucket
        - currency
        - totals
        - by_agent
        - by_provider
        - by_model
        - series
        - budgets
      properties:
        schema_version:
          type: string
          enum:
            - ai-usage-summary.v1
        tenant_id:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        bucket:
          type: string
        currency:
          type: string
        totals:
          $ref: '#/components/schemas/AiUsageTotalsV1'
        by_agent:
          type: array
          items:
            $ref: '#/components/schemas/AiUsageBreakdownV1'
        by_provider:
          type: array
          items:
            $ref: '#/components/schemas/AiUsageBreakdownV1'
        by_model:
          type: array
          items:
            $ref: '#/components/schemas/AiUsageBreakdownV1'
        series:
          type: array
          items:
            $ref: '#/components/schemas/AiUsageSeriesPointV1'
        budgets:
          type: array
          items:
            $ref: '#/components/schemas/AiBudgetLimitV1'
...
    PollenError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        trace_id:
          type: string
...
servers:
  - url: http://127.0.0.1:17891
    description: Local Control Plane / Mock Cloud
    variables: {}
  - url: https://api.pollen.cloud
    description: Pollen Cloud
    variables: {}